Refactor worker a little
